reduce races on Packages.gz on package server
When uploads happen in parallel, the likelihood of producing corrupt
Packages.gz is rather high. This can even happen within a single build.

We mitigate the problem by uploading the new packages into a unique
directory, building the new Packages.gz in this directory and moving it
and the new deb files into the final destination with a single mv command.
Not perfect, but probably good enough.